# Eliminate duplicate logical right shifts