When doing an logical right shift, there's no need to append another
zero bit logical right shift to coerce to 32 bit unsigned.