@shurcooL 2017-02-03
Interesting, thanks for the PR.

It's strange that CI isn't running on it... @neelance, any idea why? [Circle CI](https://circleci.com/gh/gopherjs/gopherjs) seems to be still on for this project.

Have you tried running any benchmarks? I understand this is meant to be an optimization, so it'd be helpful to have some numbers in order to understand the change better.

It would also be helpful to run Go1 benchmarks before and after, to see how widely applicable this optimization is. See #573, it also includes directions for running the benchmarks.