@neelance 2017-02-12
I am surprised that the JS engine does not eliminate the unnecessary right shift, it should be very easy to detect. I am also not sure if your change can replace `fixNumber` in all cases. I have to investigate it further.