@jbowes 2017-02-12
I wasn't expecting this change to have any performance impact; I was just trying to remove some generated code that didn't seem strictly necessary. It was really just something I saw while investigating #584 

Feel free to pass on this one.