@shurcooL 2017-02-13
> I wasn't expecting this change to have any performance impact; I was just trying to remove some generated code that didn't seem strictly necessary.

Ah, thanks for clarifying that. It's helpful to know what the goal is.