@neelance 2017-02-16
Yeah, I'd like to pass on this one. I'm not 100% sure that it is correct in all circumstances and I don't have the time to investigate much more. Thanks for the PR anyways. :-)