@neelance 2017-02-16
@shurcooL Can you really confirm a positive performance impact? Did all tests pass for you? (unfortunately CI skipped this PR)