@shurcooL 2017-02-16
What do you mean by confirm a positive performance impact?

I can confirm the comment I made is an accurate representation of my test run. Did you want me to run the benchmarks an extra time to confirm?

I hadn't run the tests locally yet.